+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/FizzBuzzType.java
@@ -1,4 +1,4 @@
-package helloworld;
+package fizzbuzz;
 
 public enum FizzBuzzType {
     one(new FizzBuzzType01()),
