+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/FizzBuzzType02.java
@@ -1,4 +1,4 @@
-package helloworld;
+package fizzbuzz;
 
 public class FizzBuzzType02 extends IFizzBuzzType {
     @Override
