+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/FizzBuzzValueCommand.java
@@ -1,4 +1,4 @@
-package helloworld;
+package fizzbuzz;
 
 public class FizzBuzzValueCommand implements IFizzBuzzCommand{
     private FizzBuzzType _type;
