+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/FizzBuzzValuesCommand.java
@@ -1,4 +1,4 @@
-package helloworld;
+package fizzbuzz;
 
 
 import java.util.ArrayList;
