+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/IFizzBuzzCommand.java
@@ -1,4 +1,4 @@
-package helloworld;
+package fizzbuzz;
 
 public interface IFizzBuzzCommand {
     void execute(int arg);
