+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/IFizzBuzzType.java
@@ -1,4 +1,4 @@
-package helloworld;
+package fizzbuzz;
 
 public abstract class IFizzBuzzType {
     public abstract FizzBuzzValue generate(int number);
