+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/test/java/fizzbuzz/FizzBuzzTest.java
@@ -1,5 +1,5 @@
 
-package helloworld;
+package fizzbuzz;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
