+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/test/java/helloworld/AppTest.java
@@ -3,6 +3,7 @@ package helloworld;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+
 import org.junit.Test;
 
 public class AppTest {
